Add protothread_init() and protothread_deinit()
These two functions allow static creation of the protothread
state.  protothread_deinit() is not strictly necessary but it
includes the various debug checks that should hold true in the
case where we have finished all threads.  It is also a good
idea from a stable API standpoint to have a deinitialization
function for future usage.

protothread_create() and protothread_free() were rewritten as
wrappers around protothread_init() and protothread_deinit().