Adding pt_reset to reset threads to initial state
This basically makes the next pt_resume fall through (like it
does in the initial case).  This can be used to reset threads or
thread functions in interrupt conditions.

It can also be used for resets within the thread function
itself:

static pt_t
my_thread(void * const env)
{
    if (has_interrupt()) {
        pt_reset(s) ;
    }

    pt_resume(s) ;
    [ ... ]
    return PT_DONE ;
}

Obviously the thread in question should be written to expect
this kind of operation.

Added a test to ensure the proper operation.